fix(compiler): QueryResultMatch wrap/cow match StableNodeRef widths (#3990, HIGH)
Widen wrap_epoch to uint32 and cow_epoch_at_capture to uint64 so a held
schema-2 QueryResult cannot alias after wrap 65536. Production freshness
compares wrap without the uint16 cast (captured 0 vs live 65536 is stale,
same as is_valid). Soft wrap=0 skip unchanged. No new query key.

+++ scripts/check_restamp_budget_hard_gate.py
@@ -238,8 +238,8 @@ def _self_test() -> int:
         std::uint32_t tenant_id = 0;
         std::uint32_t fiber_id = 0;
         std::uint32_t mutation_id_at_capture = 0;
-        std::uint16_t wrap_epoch = 0;
-        std::uint16_t cow_epoch_at_capture = 0;
+        std::uint32_t wrap_epoch = 0;
+        std::uint64_t cow_epoch_at_capture = 0;
         std::uint8_t boundary_pinned = 0;
         std::uint8_t reserved = 0;
         constexpr bool has_full_provenance() const noexcept;
